Moved file upload functionality back into the upload component. Directory now automatically updated on file upload.
